Move testing stuff to a Debug Feature
